Fix calculations that may lose precision compared to their target type
PiperOrigin-RevId: 444861268

+++ library/extractor/src/main/java/com/google/android/exoplayer2/extractor/ts/AdtsExtractor.java
@@ -363,6 +363,7 @@ public final class AdtsExtractor implements Extractor {
    * @return The stream bitrate.
    */
   private static int getBitrateFromFrameSize(int frameSize, long durationUsPerFrame) {
-    return (int) ((frameSize * C.BITS_PER_BYTE * C.MICROS_PER_SECOND) / durationUsPerFrame);
+    return (int)
+        ((frameSize * ((long) C.BITS_PER_BYTE) * C.MICROS_PER_SECOND) / durationUsPerFrame);
   }
 }
